make nonlinear static method with functors to take any linear solve routine
