+++ include/linalg/_linalg.h
@@ -1,0 +1,4 @@
+#include "bsr_mat.h"
+#include "dense_mat.h"
+#include "precond.h"
+#include "vec.h"
